+++ app/src/main/java/org/sopt/at/presentation/home/model/SubBannerItem.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.presentation.home.data
+package org.sopt.at.presentation.home.model
 
 data class SubBannerItem(
     val rank : Int,
