+++ app/src/main/java/org/sopt/at/presentation/home/model/TabAssets.kt
@@ -1,0 +1,10 @@
+package org.sopt.at.presentation.home.model
+
+import kotlinx.collections.immutable.ImmutableList
+
+data class TabAssets(
+    val rankedTitle: String,
+    val unrankedTitle: String,
+    val mainBanner: ImmutableList<MainBannerItem>,
+    val subBanner: ImmutableList<SubBannerItem>
+)
